+++ zingo-cli/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Deprecated
 
 ### Added
+`--seed-derivation` option - lets the user select the seed-derivation scheme when the user makes a wallet.
 
 ### Changed
 `remove_transaction` command - now only allows transactions with the new `Failed` status to be removed.